+++ bundles/aero.minova.rcp.constant/src/aero/minova/rcp/constants/Constants.java
@@ -113,6 +113,10 @@ public class Constants {
 	 * Benutzt mit einem String, öffnet einen einfachen Fehler-Dialog
 	 */
 	public static final String BROKER_SHOWERRORMESSAGE = "aero/minova/rcp/ShowErrorMessage";
+	/**
+	 * Benutzt mit einem String, öffnet einen einfachen Information-Dialog
+	 */
+	public static final String BROKER_SHOWINFORMATIONMESSAGE = "aero/minova/rcp/ShowInformationMessage";
 
 	public static final String BROKER_SELECTSTATISTIC = "aero/minova/rcp/SelectStatistic";
 	public static final String BROKER_SENDEVENTTOHELPER = "aero/minova/rcp/SendEventToHelper";